data_pipeline/requirements.txt
requests
psycopg2-binary
python-dotenv
sqlalchemy

